utils/pyproject.toml
[tool.poetry]
name = "big-button-printer-server"
version = "0.1.0"
description = ""
authors = ["Quentin Suire <dev@quentinsuire.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"
click = "^8.0.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
